refactor(nicknamer): move `names` module to top-level
Relocated the `names` module from `commands` to the top-level for better organization and alignment with streamlined architecture. Updated all imports and references to maintain consistency across the codebase.
