+++ nicknamer/src/nicknamer/commands/mod.rs
@@ -1,9 +1,8 @@
 use crate::nicknamer::connectors::discord;
 use crate::nicknamer::connectors::discord::server_member;
+use crate::nicknamer::names;
 use std::fmt::{Display, Formatter};
 use thiserror::Error;
-
-pub(crate) mod names;
 
 #[derive(Debug, PartialEq, Default)]
 pub struct User {
